feat: allow to import local models (#771)
* feat: allow to import local models


* fix: add tests


* fix: skip inference server which are using a model which do not exist on the catalog


* fix: create disposable when adding catalogUpdate listener and dispose it when modelManager is disposed


* fix: show inference servers if they have no models, display error if something unexpected happen when adding a new model and fix css inputs


* fix: add size and creation date and skip settimeout for specific channels


* fix: add tests


* fix: move notimeoutchannels to external file and remove memory prop


* Update packages/frontend/src/pages/ImportModels.svelte


* Update packages/frontend/src/pages/ImportModels.svelte


* Update packages/frontend/src/lib/table/model/ModelColumnName.svelte


* fix: fix format


* fix: show N/A when memory value is not available


* fix: set memory equals to file size


* fix: fix tests


---------

+++ packages/shared/src/messages/MessageProxy.ts
@@ -18,6 +18,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import type { Webview } from '@podman-desktop/api';
+import { noTimeoutChannels } from './NoTimeoutChannels';
 
 export interface IMessage {
   id: number;
@@ -198,12 +199,14 @@ export class RpcBrowser {
     } as IMessageRequest);
 
     // Add some timeout
-    setTimeout(() => {
-      const { reject } = this.promises.get(requestId) || {};
-      if (!reject) return;
-      reject(new Error('Timeout'));
-      this.promises.delete(requestId);
-    }, 5000);
+    if (!noTimeoutChannels.includes(channel)) {
+      setTimeout(() => {
+        const { reject } = this.promises.get(requestId) || {};
+        if (!reject) return;
+        reject(new Error('Timeout'));
+        this.promises.delete(requestId);
+      }, 5000);
+    }
 
     // Create a Promise
     return promise;
